Plotting additions

+++ FBI/plotting/plot_main.py
@@ -2,8 +2,9 @@ import apexpy
 import matplotlib.pyplot as plt
 import datetime as dt
 from os import path as pathy
-from FBI.plotting.axis import get_local_axis
-from FBI.plotting.plot import plot_noon_line, plot_vecs_model_darn_grid, plot_potential_contours, plot_data_locs
+from FBI.plotting.axis import get_local_axis, get_polar_axis
+from FBI.plotting.plot import plot_noon_line, plot_vecs_model_darn_grid, plot_potential_contours, plot_data_locs, \
+    plot_boundary_box
 
 
 def lompe_scan_plot_vectors(path, lompe):
@@ -72,6 +73,44 @@ def lompe_scan_plot_potential(path, lompe):
 
         # Locations of SuperDARN data
         plot_data_locs(lompe, ax, apex=None, time=None, coord=coord)
+
+        plt.savefig(save_path, dpi=300, bbox_inches='tight')
+        plt.close('all')
+    else:
+        print('Allready processed: ' + save_path)
+
+
+def lompe_scan_plot_potential_polar(path, lompe):
+    """
+
+    :param path:
+    :param lompe:
+    :return:
+    """
+
+    plt.rcParams['text.usetex'] = True
+
+    scan_time = dt.datetime(lompe['scan_year'][0], lompe['scan_month'][0], lompe['scan_day'][0], lompe['scan_hour'][0],
+                            lompe['scan_minute'][0], lompe['scan_second'][0], lompe['scan_millisec'][0])
+    save_path = path + scan_time.strftime("polar_pot_%Y-%m-%d %H%M%S") + '.png'
+
+    # Check plot doesn't already exist
+    if pathy.isfile(save_path) is False:
+        # Apex coordinate stuff
+        apex = apexpy.Apex(scan_time, refh=300)
+
+        # Global polar axis
+        ax, coord = get_polar_axis(scan_time, apex)
+        plt.title(scan_time.strftime("%Y-%m-%d %H:%M:%S"))
+
+        # Electric potentials
+        plot_potential_contours(lompe, ax, apex, scan_time, coord='mlt')
+
+        # Locations of SuperDARN data
+        plot_data_locs(lompe, ax, apex=apex, time=scan_time, coord=coord)
+
+        # Boundary box
+        plot_boundary_box(lompe, ax, apex, scan_time)
 
         plt.savefig(save_path, dpi=300, bbox_inches='tight')
         plt.close('all')
